+++ tests/geometry/test_inscribed_rect.py
@@ -23,7 +23,7 @@ import numpy as np
 import pytest
 
 from bluemira.display.plotter import PlotOptions, plot_2d
-from bluemira.geometry._deprecated_tools import make_circle_arc
+from bluemira.geometry._private_tools import make_circle_arc
 from bluemira.geometry.coordinates import Coordinates, get_area, in_polygon
 from bluemira.geometry.face import BluemiraFace
 from bluemira.geometry.inscribed_rect import _rect, inscribed_rect_in_poly
